More automake(1) integration.
The version-independent installation directories bin/, doc/, include/,
lib/, and man/ are now in the top-level source-directory.  This was a
necessary compromise in order to get automake(1) to work with the LDM
"runtime" switch.

The "etc/" installation-directory is special.  For the maintainer,
Steve, it's also in the top-level source-directory, but for everyone
else, it's in the LDM user's home directory.