v6.9.0.10
Improved output redirection of su(1) and sudo(1) in configure(1) script
and makefile so that the user will get any error messages even if they
redirect output to a file.

Modified plotMetrics: ported to use regutil(1) and added test for
non-existent data-files.

Improved locking in ldmadmin(1): replaced lock-file existence with
non-blocking call to flock().