Improved memory-mapping in pq(3) library
Changed the build-procedure for the product-queue library (pq(3))
to improve throughput.  Before, the build-procedure was configured
to create a pq(3) library that always memory-mapped the
data-products individually (rather than memory-mapping the entire
product-queue) if the C compiler was c89(1) or c99(1) and the
"int", "long", and pointer types were 32-bits but the "off_t" type
was at least 64-bits. Now, the pq(3) library will always memory-map
the product-queue in a single call if possible regardless of the
programming environment. This reduces the number of calls to the
system mmap(2) and munmap(2) functions and, consequently, increases
throughput.