Logging and documentation changes
Demoted the logging-level of the minimum virtual residence time (MVRT)
data-product message from NOTICE to INFO in the product-queue library so
that it will only be emitted if verbose logging is enabled.

Removed all attributes from <body> elements in the HTML documentation to
enable correct working with the new UPC website.