Changed subscription-modification by ULDB
Now, the ULDB only modifies the subscription by a downstream LDM if the
transfer mode is primary. This is done because the ULDB feature
interacts with the autoswitching feature. An autoswitch from alternate
mode to primary mode risks receiving many products that won't make it
into the product-queue because they've already been received due to the
delay in termination of the (previous) upstream LDM in alternate mode.