Improved ULDB and its testing.
Daryl Herzmann's LDM received a SIGSEGV in the ULDB module (ticket
JVF-237504). Consequently, the unit-test of that module was improved and
a robustness test was added. Also, blocking of signals was added to the
public functions of the ULDB module that modified the database.

Unfortunately, none of these changes appear to have been necessary: the
ULDB module passes the improved unit-test even without the
signal-blocking changes.