Removed use of "source" information in parsing configuration-file.
"source" is the pathname and line number of the configuration-file. The
problem was that such information could be obsolete due to buffering by
the scanner and parser when it was used -- potentially leading to a
segmentation fault when attempting to duplicate a NULL pathname with
which to tag a REQUEST entry.