Reduced the delivery latency of products.
For products with relatively large inter-product temporal gaps. Modified the
RPC calls that LDM-6 makes to send data products in primary mode by switching
from "batched" RPC calls to asynchronous "message-passing" RPC calls.
Specifically, the XDR function to decode the (non-existent) response was
changed from `NULL` to `xdr_void` in the functions `notification_6`,
`hereis_6`, and `blkdata_6`. The results are highly dependent on the
characteristics of the data flow; consequently, differences might not be
noticed.