Added SIGHUP handler used to reopen log file.
Log files created using noaaportIngester's "-l" option can now be
rotated without completely killing or restarting the noaaportIngester
process. This allows a user to mimic behavior already offered by
(r)syslog (when using the "-u" option).

This code does not actually rename or modify any files, it simply closes
the existing descriptor to a log file that may have been specified when
the program was executed. Any further log messages will result in
opening a new file descriptor to the original filename, creating a new
file if the original was moved, or simply reopening the original file if
it was not moved.

An expected workflow using this new functionality might look like this:
1) Rename logfile to logfile.1
2) Logging continues in logfile.1 because a file descriptor to the
   original inode is still open
3) Send SIGHUP to noaaportIngester process, thereby closing the fd to
   logfile.1 and reopening a (new) file named logfile, which is now the
   destination of all log messages